requirements.txt
sqlalchemy
psycopg2-binary
beautifulsoup4==4.10.0
psycopg2-binary==2.9.1
pandas==1.4.0
requests==2.26.0
lxml==4.7.3

